+++ pkg/knative/commands/handler_tail_test.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package commands_knative_test
+package knative_commands_test
 
 import (
 	"context"
@@ -23,7 +23,7 @@ import (
 	"time"
 
 	"github.com/projectriff/cli/pkg/cli"
-	commands_knative "github.com/projectriff/cli/pkg/riff/commands/knative"
+	knativecommands "github.com/projectriff/cli/pkg/knative/commands"
 	rifftesting "github.com/projectriff/cli/pkg/testing"
 	kailtesting "github.com/projectriff/cli/pkg/testing/kail"
 	requestv1alpha1 "github.com/projectriff/system/pkg/apis/knative/v1alpha1"
@@ -36,21 +36,21 @@ func TestHandlerTailOptions(t *testing.T) {
 	table := rifftesting.OptionsTable{
 		{
 			Name: "invalid resource",
-			Options: &commands_knative.HandlerTailOptions{
+			Options: &knativecommands.HandlerTailOptions{
 				ResourceOptions: rifftesting.InvalidResourceOptions,
 			},
 			ExpectFieldError: rifftesting.InvalidResourceOptionsFieldError,
 		},
 		{
 			Name: "valid resource",
-			Options: &commands_knative.HandlerTailOptions{
+			Options: &knativecommands.HandlerTailOptions{
 				ResourceOptions: rifftesting.ValidResourceOptions,
 			},
 			ShouldValidate: true,
 		},
 		{
 			Name: "since duration",
-			Options: &commands_knative.HandlerTailOptions{
+			Options: &knativecommands.HandlerTailOptions{
 				ResourceOptions: rifftesting.ValidResourceOptions,
 				Since:           "1m",
 			},
@@ -58,7 +58,7 @@ func TestHandlerTailOptions(t *testing.T) {
 		},
 		{
 			Name: "invalid duration",
-			Options: &commands_knative.HandlerTailOptions{
+			Options: &knativecommands.HandlerTailOptions{
 				ResourceOptions: rifftesting.ValidResourceOptions,
 				Since:           "1",
 			},
@@ -157,5 +157,5 @@ func TestHandlerTailCommand(t *testing.T) {
 		},
 	}
 
-	table.Run(t, commands_knative.NewHandlerTailCommand)
+	table.Run(t, knativecommands.NewHandlerTailCommand)
 }
